+++ src/intersection-root-context.js
@@ -1,0 +1,7 @@
+// @flow
+import * as React from 'react';
+
+export const IntersectionRootContext: React$Context<any> = React.createContext({
+  observe: null,
+  unobserve: null
+});